fix: selection highlight on last line and add Execute All shortcut (#770)
