+++ docs/features/keyboard-shortcuts.mdx
@@ -36,6 +36,7 @@ TablePro is keyboard-driven. Most actions have shortcuts, and most menu shortcut
 | Action | Shortcut | Description |
 |--------|----------|-------------|
 | Execute query | `Cmd+Enter` | Run query at cursor, or all selected statements sequentially |
+| Execute all statements | `Cmd+Shift+Enter` | Run all statements in the editor sequentially |
 | Cancel query | `Cmd+.` | Stop the currently running query |
 | Explain query | `Option+Cmd+E` | Show execution plan for query at cursor |
 | Format SQL | `Cmd+Shift+L` | Format SQL query |
